requirements.txt
PyYAML==5.4.1
selenium==4.1.5
